Move OrbitEventsSection into page subfolders
