+++ src/app/orbit/page.tsx
@@ -6,7 +6,6 @@ import { CTASection } from '@/components/CTASection'
 import { FocusAreaCard } from '@/components/FocusAreaCard'
 import { HomeExploreSectionCard } from '@/components/HomeExploreSectionCard'
 import { OrbitAmbassadorCard } from '@/components/OrbitAmbassadorCard'
-import { OrbitEventsSection } from '@/components/OrbitEventsSection'
 import { PageHeader } from '@/components/PageHeader'
 import { PageLayout } from '@/components/PageLayout'
 import { PageSection } from '@/components/PageSection'
@@ -24,6 +23,7 @@ import { PATHS } from '@/constants/paths'
 import { FILECOIN_FOUNDATION_URLS } from '@/constants/siteMetadata'
 import { graphicsData } from '@/data/graphicsData'
 
+import { OrbitEventsSection } from './components/EventsSection'
 import { ambassadorsData } from './data/ambassadorsData'
 import { exploreOrbitData } from './data/exploreOrbitData'
 import { programFeaturesAndPerksData } from './data/programFeaturesAndPerksData'
